fix(#110): move mixed_forward_batch_dispatch to own file (LoC cap) + update arena alloc-count test
decode_b.rs exceeded the 500-LoC cap; relocate the #110 dispatch into
trait_impl/mixed_batch.rs. Update buffers::tests::test_buffer_arena_alloc
expected alloc count 19 -> 20 for the new decode_logits_staging buffer.
